Auth API removed from repo. GodUser added.
