+++ .gitignore
@@ -1,4 +1,2 @@
 node_modules/
-output/*
-!output/.gitkeep
 .env
