--- api-logic/.deployment
@@ -1,2 +1,0 @@
-[config]
-SCM_DO_BUILD_DURING_DEPLOYMENT=true
