--- api-logic/Dockerfile
@@ -1,11 +1,0 @@
-FROM node:10
-
-ADD . /app
-
-WORKDIR /app
-
-RUN npm install
-
-USER node
-
-CMD [ "./cmd.sh" ]
