--- api-logic/src/api/DummyLogger.js
@@ -1,16 +1,0 @@
-
-class DummyLogger {
-  logEvent (name, properties) {
-    console.log({ name, properties });
-  }
-
-  logException (error) {
-    console.error(error);
-  }
-
-  logTrace (message) {
-    console.log(message);
-  }
-}
-
-module.exports = DummyLogger;
