--- api-logic/src/api/ITokenValidator.js
@@ -1,5 +1,0 @@
-const Interface = require('@amadek/js-sdk/Interface');
-
-const ITokenValidator = new Interface('ITokenValidator', [ 'validate' ], []);
-
-module.exports = ITokenValidator;
