--- api-logic/src/logic/IDbConnector.js
@@ -1,5 +1,0 @@
-const Interface = require('@amadek/js-sdk/Interface');
-
-const IDbConnector = new Interface('IDbConnector', ['connect', 'close'], []);
-
-module.exports = IDbConnector;
