--- api-logic/src/logic/IFuelConsumption.js
@@ -1,5 +1,0 @@
-const Interface = require('@amadek/js-sdk/Interface');
-
-const IFuelConsumption = new Interface('IFuelConsumption', [], ['liters', 'kilometers', 'fuelPrice', 'created']);
-
-module.exports = IFuelConsumption;
