+++ cmd.sh
