+++ docker-compose.yml
@@ -1,24 +1,11 @@
-fuel-cost-api-auth:
-  container_name: fuel-cost-api-auth
-  build: ./api-auth
-  net: fuel-cost-net
-  ports:
-    - '8181:8181'
-  volumes:
-    - ./api-auth:/app
-  environment:
-    - NODE_ENV=development
-    - PORT=8181
-    - PUT_TOKEN_URL=http://fuel-cost-api:3000/token
-
 fuel-cost-api:
   container_name: fuel-cost-api
-  build: ./api-logic/
+  build: ./
   net: fuel-cost-net
   ports:
     - '3000:3000'
   volumes:
-    - ./api-logic:/app
+    - ./:/app
   environment:
     - NODE_ENV=development
     - PORT=3000
