+++ src/api/DummyLogger.js
@@ -1,0 +1,16 @@
+
+class DummyLogger {
+  logEvent (name, properties) {
+    console.log({ name, properties });
+  }
+
+  logException (error) {
+    console.error(error);
+  }
+
+  logTrace (message) {
+    console.log(message);
+  }
+}
+
+module.exports = DummyLogger;
